feature: Integrate alerts and dynamic move display
All console.logs() have been replaced with the AlertSystem. The
AlertSystem no longer takes a domSelector because it will always use
main-view.

Entities will refer to their moves based on an index number instead of
a name. The data will be validated before it comes in so it will always
be valid. They will have a movesList passed into them that they can
use moves from.

Moves will be dynamically loaded based on the movesList based in.